requirements.txt
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
pycparser==2.21
PyJWT==2.3.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
